refactor faucet into `useFaucet`
